pyproject.toml
[project]
name = "llm-rpg"
version = "0.1.0"
description = ""
authors = [{ name = "WoutVossen", email = "vossen.w@hotmail.com" }]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai>=1.57.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "ollama>=0.5.1",
    "pygame>=2.6.1",
    "unfake>=1.0.7",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "diffusers>=0.36.0",
    "transformers>=4.57.3",
    "tokenizers>=0.22.1",
    "peft>=0.18.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pre-commit>=4.1.0",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_rpg"]
include = ["src/llm_rpg/assets/**/*"]

[tool.hatch.build.targets.sdist]
include = ["src/llm_rpg/assets/**/*"]
